算法最终版

+++ job.c
@@ -14,8 +14,12 @@ int jobid=0;
 int siginfo=1;
 int fifo;
 int globalfd;
+int changleable;
 
 struct waitqueue *head=NULL;
+struct waitqueue *head2=NULL;
+struct waitqueue *head5=NULL;
+
 struct waitqueue *next=NULL,*current =NULL;
 
 /* 调度程序 */
@@ -69,13 +73,19 @@ void updateall()
 	struct waitqueue *p;
 
 	/* 更新作业运行时间 */
-	if(current)
+	if(current){
 		current->job->run_time += 1; /* 加1代表1000ms */
+        changleable=0;
+        if(current->job->fromqueue==1&&current->job->run_time==1||
+           current->job->fromqueue==2&&current->job->run_time==2||
+           current->job->fromqueue==5&&current->job->run_time==5)
+            changleable=1;
+	}
 
 	/* 更新作业等待时间及优先级 */
 	for(p = head; p != NULL; p = p->next){
 		p->job->wait_time += 1000;
-		if(p->job->wait_time >= 5000 && p->job->curpri < 3){
+		if(p->job->wait_time >= 10000 /*&& p->job->curpri < 3*/){
 			p->job->curpri++;
 			p->job->wait_time = 0;
 		}
@@ -89,18 +99,45 @@ struct waitqueue* jobselect()
 
 	select = NULL;
 	selectprev = NULL;
-	if(head){
-		/* 遍历等待队列中的作业，找到优先级最高的作业 */
-		for(prev = head, p = head; p != NULL; prev = p,p = p->next)
-			if(p->job->curpri > highest){
-				select = p;
-				selectprev = prev;
-				highest = p->job->curpri;
-			}
-			selectprev->next = select->next;
-			if (select == selectprev)
-				head = NULL;
-	}
+     /* 遍历等待队列中的作业，找到优先级最高的作业 */
+    if(head){
+        for(prev = head, p = head; p != NULL; prev = p,p = p->next)
+            if(p->job->curpri > highest){
+                select = p;
+                select->job->fromqueue = 1;
+                selectprev = prev;
+                highest = p->job->curpri;
+            }
+            selectprev->next = select->next;
+            if (select == selectprev)
+                head = NULL;
+
+    }
+    else if(head2){
+        for(prev = head2, p = head2; p != NULL; prev = p,p = p->next)
+            if(p->job->curpri > highest){
+                select = p;
+                select->job->fromqueue = 2;
+                selectprev = prev;
+                highest = p->job->curpri;
+            }
+            selectprev->next = select->next;
+            if (select == selectprev)
+                head2 = NULL;
+    }
+    else if(head5){
+        for(prev = head5, p = head5; p != NULL; prev = p,p = p->next)
+            if(p->job->curpri > highest){
+                select = p;
+                select->job->fromqueue = 5;
+                selectprev = prev;
+                highest = p->job->curpri;
+            }
+            selectprev->next = select->next;
+            if (select == selectprev)
+                head5 = NULL;
+    }
+
 	return select;
 }
 
@@ -123,9 +160,10 @@ void jobswitch()
 		current = NULL;
 	}
 
-	if(next == NULL && current == NULL) /* 没有作业要运行 */
-
+	if(next == NULL && current == NULL){ /* 没有作业要运行 */
+        printf("No more job to run!");
 		return;
+	}
 	else if (next != NULL && current == NULL){ /* 开始新的作业 */
 
 		printf("begin start new job\n");
@@ -135,7 +173,7 @@ void jobswitch()
 		kill(current->job->pid,SIGCONT);
 		return;
 	}
-	else if (next != NULL && current != NULL){ /* 切换作业 */
+	else if (next != NULL && current != NULL && changleable == 1){ /* 切换作业 */
 
 		printf("switch to Pid: %d\n",next->job->pid);
 		kill(current->job->pid,SIGSTOP);
@@ -144,11 +182,24 @@ void jobswitch()
 		current->job->state = READY;
 
 		/* 放回等待队列 */
-		if(head){
-			for(p = head; p->next != NULL; p = p->next);
-			p->next = current;
-		}else{
-			head = current;
+		switch(current->job->fromqueue){
+            case 1:
+                if(head2){
+                    for(p = head2; p->next != NULL; p = p->next);
+                    p->next = current;
+                }else{
+                    head2 = current;
+                }
+                break;
+            case 2:
+            case 5:
+                if(head5){
+                    for(p = head5; p->next != NULL; p = p->next);
+                    p->next = current;
+                }else{
+                    head5 = current;
+                }
+                break;
 		}
 		current = next;
 		next = NULL;
@@ -157,6 +208,7 @@ void jobswitch()
 		kill(current->job->pid,SIGCONT);
 		return;
 	}else{ /* next == NULL且current != NULL，不切换 */
+	    printf("Done!");
 		return;
 	}
 }
